# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: ResQNet Backend Core API
  version: 1.0.0
  description: |
    Mission, sector, operator, auth and audit endpoints for the ResQNet
    command platform (Smart India Hackathon Phase 1, software-first).

    **Bootstrap**

    ```
    docker compose up -d db api
    docker compose run --rm seed
    ```

    Seeded accounts all share the password `resqnet-demo`:

    | email | role | may |
    |---|---|---|
    | admin@resqnet.demo | admin | everything, plus user management |
    | operator@resqnet.demo | operator | create missions/sectors, change mission status |
    | viewer@resqnet.demo | viewer | read only |
servers:
  - url: http://localhost:3000
    description: Local Docker Compose demo
...
security:
  - bearerAuth: []
paths:
...
  /missions:
    post:
      tags: [missions]
      summary: Create a mission and assign its search sectors
      description: |
        Demo steps 1–2 of Section 27. The zone is split into `sector_count`
        equal strips along its longer axis, so a new mission always arrives
        with a usable sector layout. The mission starts in `created`; nothing
        flies until it is moved to `active`.

        Requires `mission:create` (operator or admin).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMission" }
      responses:
        "201":
          description: Mission created, with its generated sectors
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Mission"
                  - type: object
                    properties:
                      sectors:
                        type: array
                        items: { $ref: "#/components/schemas/Sector" }
        "400":
          {
            description: Malformed zone polygon,
            sector_count out of range,
            or duplicate mission_id,
          }
        "403": { description: Role lacks mission:create }
    get:
      tags: [missions]
      summary: List missions, newest first
      responses:
        "200":
          description: Missions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Mission" }
  /missions/{id}:
    get:
      tags: [missions]
      summary: Fetch one mission
      parameters:
        - $ref: "#/components/parameters/MissionId"
      responses:
        "200":
          description: The mission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Mission" }
        "404": { description: No such mission }
  /missions/{id}/status:
    patch:
      tags: [missions]
      summary: Drive the mission state machine
      description: |
        Allowed edges, and the Section 10.6 event each publishes:

        | from → to | event |
        |---|---|
        | created → active | `mission.started` |
        | paused → active | `mission.started` |
        | active → paused | `mission.paused` |
        | active/paused → completed | `mission.completed` |

        `completed` is terminal. Chirag's gateway starts and stops simulated
        drone motion on these events, so an invalid transition is rejected
        before anything is published.

        Requires `mission:update-status`.
      parameters:
        - $ref: "#/components/parameters/MissionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [created, active, paused, completed]
      responses:
        "200":
          description: Updated mission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Mission" }
        "400": { description: Transition not allowed from the current status }
        "403": { description: Role lacks mission:update-status }
        "404": { description: No such mission }
  /missions/{id}/sectors:
    get:
      tags: [sectors]
      summary: Sectors of a mission, ordered SECTOR-A first
      parameters:
        - $ref: "#/components/parameters/MissionId"
      responses:
        "200":
          description: Sectors
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Sector" }
        "404": { description: No such mission }
  /sectors:
    post:
      tags: [sectors]
      summary: Create or replace one sector definition
      description: |
        Upsert on `(mission_id, sector_id)`. Mission creation has already
        claimed SECTOR-A..N with generated strips, so pinning an exact polygon
        — the layout `apps/simulator/src/sectors.ts` flies, for instance — or
        assigning a drone overwrites the generated strip rather than colliding
        with it.

        The polygon must lie inside the mission zone. Requires `sector:create`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSector" }
      responses:
        "201":
          description: The stored sector
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sector" }
        "400":
          {
            description: Malformed polygon,
            polygon outside the mission zone,
            or bad sector_id,
          }
        "403": { description: Role lacks sector:create }
        "404": { description: No such mission }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: The `access_token` from /auth/login. Refresh tokens are rejected here.
  parameters:
    MissionId:
      name: id
      in: path
      required: true
      description: The mission's uuid or its external mission_id (`MISSION-DEMO-1`). Either resolves.
      schema: { type: string }
      example: MISSION-DEMO-1
...
  schemas:
    GeoJsonPolygon:
      type: object
      description: GeoJSON Polygon, SRID 4326, single closed exterior ring, coordinates as [lon, lat].
      required: [type, coordinates]
      properties:
        type: { type: string, enum: [Polygon] }
        coordinates:
          type: array
          items:
            type: array
            items:
              type: array
              items: { type: number }
              minItems: 2
              maxItems: 2
      example:
        type: Polygon
        coordinates:
          [
            [
              [77.20, 28.610],
              [77.22, 28.610],
              [77.22, 28.620],
              [77.20, 28.620],
              [77.20, 28.610],
            ],
          ]
...
    CreateMission:
      type: object
      required: [name, zone_polygon, sector_count]
      properties:
        name:
          {
            type: string,
            maxLength: 120,
            example: Yamuna Flood Plain — Demo Sweep,
          }
        zone_polygon: { $ref: "#/components/schemas/GeoJsonPolygon" }
        sector_count:
          type: integer
          minimum: 1
          maximum: 26
          description: Capped at 26 — the SECTOR-[A-Z] ceiling the frozen telemetry contract imposes.
          example: 3
        mission_id:
          type: string
          pattern: "^MISSION-[A-Z0-9-]+$"
          description: Optional. Generated when omitted; the seed pins MISSION-DEMO-1.
    Mission:
      type: object
      properties:
        id: { type: string, format: uuid, description: Surrogate key. }
        mission_id:
          type: string
          description: External identifier every other service puts on the wire.
          example: MISSION-DEMO-1
        name: { type: string }
        status: { type: string, enum: [created, active, paused, completed] }
        zone_polygon: { $ref: '#/components/schemas/GeoJsonPolygon' }
        sector_count: { type: integer }
        created_by: { type: [string, 'null'], format: uuid }
        started_at:
          type: [string, 'null']
          format: date-time
          description: Set on the first activation only — the PRD Section 15 metric measures from mission start, not from the last resume.
        completed_at: { type: [string, 'null'], format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateSector:
      type: object
      required: [mission_id, sector_id, polygon]
      properties:
        mission_id:
          {
            type: string,
            description: The mission's uuid or its external mission_id.,
          }
        sector_id:
          type: string
          pattern: "^SECTOR-[A-Z]$"
          example: SECTOR-A
        polygon: { $ref: "#/components/schemas/GeoJsonPolygon" }
        assigned_drone_id:
          type: string
          pattern: "^DRONE-[0-9]{2}$"
          description: Plain identifier, not a foreign key — the drones table belongs to Chirag and may not exist when this service runs alone.
          example: DRONE-01
    Sector:
      type: object
      properties:
        id: { type: string, format: uuid }
        mission_id: { type: string, example: MISSION-DEMO-1 }
        sector_id: { type: string, example: SECTOR-A }
        polygon: { $ref: '#/components/schemas/GeoJsonPolygon' }
        assigned_drone_id: { type: [string, 'null'], example: DRONE-01 }
        created_at: { type: string, format: date-time }
...